some minor fixes

+++ src/components/ProjectDetails/ProjectDetail.tsx
@@ -14,6 +14,7 @@ import OperationLocationTab from './OperationLocationTab';
 import ProjectAgreementTab from './ProjectAgreementTab';
 import CostEstimateTab from './CostEstimateTab';
 import ConsumerCommitteeTab from './ConsumerCommitteeTab';
+import Swal from 'sweetalert2';
 
 
 interface FormRow {
@@ -474,6 +475,18 @@ const ProjectDetail: React.FC<ProjectDetailProps> = ({ project, onBack }) => {
   };
 
   const handleDeleteDocument = async (id: number) => {
+  const result = await Swal.fire({
+    title: 'के तपाईं यो कागजात मेटाउन निश्चित हुनुहुन्छ?',
+    text: 'यो कार्य फिर्ता गर्न सकिँदैन!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#3085d6',
+    confirmButtonText: 'हो, मेटाउनुहोस्',
+    cancelButtonText: 'रद्द गर्नुहोस्',
+  });
+
+  if (result.isConfirmed) {
     try {
       const token = localStorage.getItem('access_token');
       await axios.delete(`http://213.199.53.33:8000/api/projects/${project.serial_number}/documents/${id}/`, {
@@ -485,8 +498,10 @@ const ProjectDetail: React.FC<ProjectDetailProps> = ({ project, onBack }) => {
       toast.success('सफलतापूर्वक मेटाइयो');
     } catch (error) {
       console.error('error deleting');
+      toast.error('मेटाउन सकिएन');
     }
-  };
+  }
+};
 
   const handleSaveDocument = (data: any) => {
     if (documentDetail) {
@@ -497,7 +512,15 @@ const ProjectDetail: React.FC<ProjectDetailProps> = ({ project, onBack }) => {
   };
 
   const handleDelete = async (id: number, type: 'official' | 'monitoring' | 'document') => {
-    if (window.confirm('के तपाईं यो मेटाउन चाहनुहुन्छ?')) {
+    const result = await Swal.fire({
+      title: 'के तपाईं यो मेटाउन चाहनुहुन्छ?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'हो, मेटाउनुहोस्',
+      cancelButtonText: 'रद्द गर्नुहोस्',
+    });
+
+    if (result.isConfirmed) {
       try {
         switch (type) {
           case 'official':
